# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduConnect API",
        "description": "Tài liệu API cho hệ thống EduConnect - Quản lý giáo dục trực tuyến",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.toanrobert.online",
            "description": "Production server"
        }
    ],
    "paths": {
        "/api/test": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Test API",
                "description": "Trả về chuỗi test đơn giản",
                "operationId": "af4eef7c415174b57a22c482494a6a28",
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng ký tài khoản mới",
                "description": "Đăng ký tài khoản người dùng mới. Sau khi đăng ký thành công, email xác thực sẽ được gửi đến địa chỉ email đã cung cấp.",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Tên người dùng (2-100 ký tự)",
                                        "type": "string",
                                        "example": "Nguyễn Văn A"
                                    },
                                    "email": {
                                        "description": "Địa chỉ email hợp lệ",
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "description": "Mật khẩu (tối thiểu 8 ký tự, bao gồm chữ hoa, chữ thường và số)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password@123"
                                    },
                                    "password_confirmation": {
                                        "description": "Xác nhận mật khẩu (phải khớp với password)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password@123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đăng ký thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registration successful. Please check your email for verification."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "user@example.com"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "UNVERIFIED"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Lỗi xác thực dữ liệu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Email đã tồn tại."
                                                    }
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Mật khẩu phải có ít nhất 8 ký tự."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Quá nhiều yêu cầu (rate limit)"
                    }
                }
            }
        },
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Danh sách người dùng (có phân trang, role, permissions, profile)",
                "description": "Trả về danh sách người dùng dạng phân trang cùng với thông tin chi tiết (profile, roles, permissions).",
                "operationId": "7682ed8f32d8216884bcdfd052b767cd",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Số trang hiện tại",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách người dùng thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "admin@educonnect.com"
                                                    },
                                                    "profile": {
                                                        "properties": {
                                                            "full_name": {
                                                                "type": "string",
                                                                "example": "Admin User"
                                                            },
                                                            "phone_number": {
                                                                "type": "string",
                                                                "example": "385-678-0558"
                                                            },
                                                            "birthday": {
                                                                "type": "string",
                                                                "example": "2001-03-26"
                                                            },
                                                            "gender": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "address": {
                                                                "type": "string",
                                                                "example": "42827 Nannie Loaf, FL 79652-7986"
                                                            },
                                                            "avatar": {
                                                                "type": "string",
                                                                "example": "avatars/default.png"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "roles": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "admin"
                                                        }
                                                    },
                                                    "permissions": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "manage_users"
                                                        }
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-10-02T10:27:06.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2025-10-02T10:27:06.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/admin/users?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/admin/users?page=44"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/admin/users?page=2"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 44
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/admin/users"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 646
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Không có quyền truy cập hoặc chưa đăng nhập"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Test",
            "description": "Test"
        },
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}